# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_route-fix-backup-20250603_161334_handlers.LogEntry:
    properties:
      data:
        additionalProperties: true
        type: object
      error:
        additionalProperties: true
        type: object
      level:
        type: string
      message:
        type: string
      method:
        type: string
      path:
        type: string
      request_id:
        type: string
      source:
        type: string
      status_code:
        type: integer
      timestamp:
        type: string
      url:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    required:
    - level
    - message
    type: object
...
  models.Media:
    properties:
      created_at:
        type: string
      description:
        description: Optional description
        type: string
      filename:
        description: Original filename from upload
        type: string
      id:
        type: string
      mime_type:
        description: Detected MIME type (e.g., "image/jpeg", "application/pdf")
        type: string
      size:
        description: File size in bytes
        type: integer
      updated_at:
        type: string
      uploaded_by:
        description: User who uploaded the file (nullable for system uploads)
        type: string
      url:
        description: Publicly accessible URL (or path to serve from)
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties: true
        type: object
      error:
        example: Invalid request
        type: string
      message:
        example: Validation failed
        type: string
    type: object
...
host: reksolindo.opuschamber.com
info:
  contact: {}
  description: Multi-tenant enterprise asset management system
  title: Reksolindo Enterprise Asset Management API
  version: "1.0"
paths:
  /api/v1/logs:
    post:
      consumes:
      - application/json
      description: Receives and processes logs from the frontend application
      parameters:
      - description: Log entry from frontend
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.LogEntry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Handle frontend logs
      tags:
      - logs
  /api/v1/logs/batch:
    post:
      consumes:
      - application/json
      description: Receives and processes multiple logs from the frontend application
      parameters:
      - description: Array of log entries from frontend
        in: body
        name: logs
        required: true
        schema:
          items:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.LogEntry'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Handle batch frontend logs
      tags:
      - logs
...
  /media/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a media file record and the corresponding file(s) from
        storage.
      parameters:
      - description: Media ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted media file
        "400":
          description: Invalid media ID format
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (JWT invalid or missing)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user lacks permission)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Media record not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error during deletion
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete media file by ID
      tags:
      - Media
  /media/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a file (image, document, etc.) and stores its metadata.
        Generates thumbnails for supported image types.
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Optional description for the file
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Successfully uploaded media file and created metadata record
          schema:
            $ref: '#/definitions/models.Media'
        "400":
          description: Invalid form data or file upload error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (JWT invalid or missing)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user lacks permission)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error during file processing or saving
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a media file
      tags:
      - Media
...
schemes:
- https
- http
swagger: "2.0"